added new cards for gettiong Genes for RE, changed URL's, bug fixes
