+++ components/data/drug_matrix_transpose/index.tsx
@@ -67,7 +67,7 @@ export const DMT = MetaNode(`DMT`)
   ))
   .story(props => ({
     abstract: `The file${props.inputs && props.inputs.file.description ? ` containing ${props.inputs.file.description}` : ''} was loaded as a drug matrix transpose.`,
-    introduction: `Drug set libraries associate biomedical terms with drugs and small molecules. Drug set libraries are stored as drug matrix transposed (.DMT) files, a tab delimited file format that describes a collection of term–drug set associations.\\ref{doi:10.1093/database/baab017}`,
+    introduction: `Drug set libraries associate biomedical terms with drugs and small molecules. Drug set libraries are stored as drug matrix transposed (.DMT) files, a tab delimited file format that describes a collection of term-drug set associations.\\ref{doi:10.1093/database/baab017}`,
     legend: `A drug set library provided by the user.`,
   }))
   .build()
@@ -115,7 +115,7 @@ export const DMTUnion = MetaNode('DMTUnion')
   })
   .story(props => ({
     abstract: `All the identified drug sets were combined using the union set operation.`,
-    introduction: `Drug set libraries associate biomedical terms with drugs and small molecules. Drug set libraries are stored as drug matrix transposed (.DMT) files, a tab delimited file format that describes a collection of term–drug set associations.\\ref{doi:10.1093/database/baab017}`,
+    introduction: `Drug set libraries associate biomedical terms with drugs and small molecules. Drug set libraries are stored as drug matrix transposed (.DMT) files, a tab delimited file format that describes a collection of term-drug set associations.\\ref{doi:10.1093/database/baab017}`,
     methods: `All drugs across all drug sets in ${props.input_refs?.dmt} were aggregated to${props.output_ref}.`,
     legend: `The set of drugs appearing anywhere across all drug sets in the drug set library.`,
   }))
@@ -133,7 +133,7 @@ export const DMTIntersection = MetaNode('DMTIntersection')
   })
   .story(props => ({
     abstract: `A consensus drug set was created using the set intersection operation.`,
-    introduction: `Drug set libraries associate biomedical terms with drugs and small molecules. Drug set libraries are stored as drug matrix transposed (.DMT) files, a tab delimited file format that describes a collection of term–drug set associations.\\ref{doi:10.1093/database/baab017}`,
+    introduction: `Drug set libraries associate biomedical terms with drugs and small molecules. Drug set libraries are stored as drug matrix transposed (.DMT) files, a tab delimited file format that describes a collection of term-drug set associations.\\ref{doi:10.1093/database/baab017}`,
     methods: `Drugs appearing in all drug sets from${props.input_refs?.dmt} were taken in${props.output_ref}.`,
     legend: `The set of drugs appearing in all drug sets in the drug set library.`,
   }))
@@ -162,7 +162,7 @@ export const DMTConsensus = MetaNode('DMTConsensus')
   })
   .story(props => ({
     abstract: `A consensus drug set was created by only retaining drugs that appear in at least two sets.`,
-    introduction: `Drug set libraries associate biomedical terms with drugs and small molecules. Drug set libraries are stored as drug matrix transposed (.DMT) files, a tab delimited file format that describes a collection of term–drug set associations.\\ref{doi:10.1093/database/baab017}`,
+    introduction: `Drug set libraries associate biomedical terms with drugs and small molecules. Drug set libraries are stored as drug matrix transposed (.DMT) files, a tab delimited file format that describes a collection of term-drug set associations.\\ref{doi:10.1093/database/baab017}`,
     methods: `Drugs appearing in at least two drug sets from${props.input_refs?.dmt} were taken in${props.output_ref}.`,
     legend: `The set of drugs appearing in at least two drug sets from the drug set library.`,
   }))
@@ -246,7 +246,7 @@ export const DrugSetsToDMT = MetaNode('DrugSetsToDMT')
   })
   .story(props => ({
     abstract: `The drug sets collected were combined into one drug set library.`,
-    introduction: `Drug set libraries associate biomedical terms with drugs and small molecules. Drug set libraries are stored as drug matrix transposed (.DMT) files, a tab delimited file format that describes a collection of term–drug set associations.\\ref{doi:10.1093/database/baab017}`,
+    introduction: `Drug set libraries associate biomedical terms with drugs and small molecules. Drug set libraries are stored as drug matrix transposed (.DMT) files, a tab delimited file format that describes a collection of term-drug set associations.\\ref{doi:10.1093/database/baab017}`,
     methods: `The drug sets from${(props.input_refs?.sets as string[]).join(', ')} are combined into a drug set library${props.output_ref}.`,
     legend: `A drug set library produced by combining drug sets.`,
   }))
@@ -264,7 +264,7 @@ export const DMTConcatenate = MetaNode('DMTConcatenate')
   })
   .story(props => ({
     abstract: `The identified drug sets were combined into one drug set library.`,
-    introduction: `Drug set libraries associate biomedical terms with drugs and small molecules. Drug set libraries are stored as drug matrix transposed (.DMT) files, a tab delimited file format that describes a collection of term–drug set associations.\\ref{doi:10.1093/database/baab017}`,
+    introduction: `Drug set libraries associate biomedical terms with drugs and small molecules. Drug set libraries are stored as drug matrix transposed (.DMT) files, a tab delimited file format that describes a collection of term-drug set associations.\\ref{doi:10.1093/database/baab017}`,
     methods: `A joint drug sets library is constructed by stacking${(props.input_refs?.dmts as string[]).join(', ')}.`,
     legend: `A drug set library produced by combining multiple drug set libraries.`,
   }))
